+++ CHANGELOG.md
@@ -22,6 +22,7 @@ IMPROVEMENTS:
 BUG FIXES:
 * `resource/vcd_vapp_vm` and `datasource/vcd_vapp_vm` can report `network.X.is_primary` attribute
   incorrectly when VM is imported to Terraform and NIC indexes in vCD do not start with 0. [GH-512] 
+* Rename docs files from `.markdown` to `.html.markdown` (Add test to check file name consistency) [GH-522]
 
 DEPRECATIONS: